refactor: 파일, 코드 리팩토링
